+++ app/javascript/components/help/CreditsSection.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+export default function CreditsSection() {
+  return (
+    <div>
+      <p>
+        A Hex Too Far was designed primarily by Douglas Triggs, with a lot of input from
+        James Scott who contributed significant suggestions, and helped with brainstorming
+        and just generally being someone who was always there to bounce ideas off of.
+      </p>
+      <p>
+        A Hex Too Far was coded entirely by Douglas Triggs.
+      </p>
+      <p>
+        Playtesters:
+      </p>
+      <ul>
+        <li>James Scott</li>
+        <li>Ricardo Gomez</li>
+        <li>Draco &quot;CptClutch&quot; Iarovoi</li>
+      </ul>
+    </div>
+  );
+}